pkg/policy: Fix L7 policy merging.
The generic L7 policy merging logic was missing from the Egress case
and the Ingress case was also not completely correct.

Factor out the common code to a new mergeL4Port() helper and test for
the presence of the 'L7Proto' field rather than an empty set of L7
rules, so that conflict is detected also when the list of L7 rules is
empty, but the L7Proto is specified.

Require the L7Proto field also in unit testing.

Detect merge conflicts due to a specified 'L7Proto' even when there are no rules.

Add a L7 policy merge test case also for egress.

Print out L7 and Kafka rules on policy trace.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>